+++ internal/webhook/podmutation/webhookhandler_suite_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sidecar_test
+package podmutation_test
 
 import (
 	"context"
