+++ internal/webhook/podmutation/webhookhandler_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sidecar_test
+package podmutation_test
 
 import (
 	"context"
@@ -33,7 +33,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
 	"github.com/open-telemetry/opentelemetry-operator/internal/config"
 	"github.com/open-telemetry/opentelemetry-operator/internal/naming"
-	. "github.com/open-telemetry/opentelemetry-operator/internal/webhook/sidecar"
+	. "github.com/open-telemetry/opentelemetry-operator/internal/webhook/podmutation"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/sidecar"
 )
